extract progress formatter as util
