+++ src/svelte/FileComponent.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
 import { Menu } from 'obsidian'
 import { settings } from 'stores'
+import { extractProgressText } from '@utils/utils'
 
 import TaskItemComponent from '@svelte/TaskItemComponent.svelte'
 import type TaskTracker from '@components/TaskTracker'
@@ -48,35 +49,6 @@ const progress = (item: TaskItem) => {
         return ((item.actual / item.expected) * 100).toFixed(2)
     }
     return 0
-}
-
-const progressText = (item: TaskItem) => {
-    let { actual, expected } = item
-    if (expected > 0) {
-        let unfinished = expected - actual
-        let max = Math.max(expected, actual)
-        if (max > 10) {
-            if (unfinished > 0) {
-                return `◌ x ${unfinished} 🍅 x ${actual}`
-            } else {
-                return `🍅 x ${expected}  🥫 x ${Math.abs(unfinished)}`
-            }
-        } else {
-            if (unfinished > 0) {
-                return `${'🍅'.repeat(actual)}${'◌'.repeat(unfinished)}`
-            } else {
-                return `${'🍅'.repeat(expected)}${'🥫'.repeat(
-                    Math.abs(unfinished),
-                )}`
-            }
-        }
-    } else {
-        return actual > 10
-            ? `🍅 x ${actual}`
-            : actual > 0
-              ? `${'🍅'.repeat(actual)}`
-              : `- -`
-    }
 }
 
 const openFile = (e: MouseEvent) => {
@@ -222,7 +194,7 @@ const showTaskMenu = (task: TaskItem) => (e: MouseEvent) => {
                                 content={item.description} />
                         </div>
                         <div class="pomodoro-tasks-progress">
-                            {progressText(item)}
+                            {extractProgressText(item)}
                         </div>
                     </div>
                 {/each}
